+++ module/core/derive_tools/tests/inc/from_inner_variants_collisions.rs
@@ -25,15 +25,3 @@ pub enum GetData
 // == end of generated
 
 include!( "./only_test/from_inner_variants.rs" );
-
-// xxx2 : implement attribute `#[ from( off ) ]`
-//
-// #[ derive( Debug, PartialEq, From ) ]
-// // #[ debug ]
-// pub enum GetData< 'a, T >
-// where
-//   T : ToString + ?Sized,
-// {
-//   FromBin( &'static [ u8 ] ),
-//   FromT( &'a T ),
-// }
