+++ module/core/derive_tools/tests/inc/mod.rs
@@ -68,8 +68,14 @@ mod from_inner_multiple_test;
 mod from_inner_variants_manual;
 #[ cfg( feature = "derive_from" ) ]
 mod from_inner_variants_derive;
+
 #[ cfg( feature = "derive_from" ) ]
-mod from_inner_variants_duplicates;
+mod from_inner_variants_duplicates_all_off;
+#[ cfg( feature = "derive_from" ) ]
+mod from_inner_variants_duplicates_some_off;
+#[ cfg( feature = "derive_from" ) ]
+mod from_inner_variants_duplicates_some_off_default_off;
+
 #[ cfg( feature = "derive_from" ) ]
 mod from_inner_variants_generics;
 #[ cfg( feature = "derive_from" ) ]
